+++ src/lib_gui/qt/project_wizard/content/QtProjectWizardContentCppStandard.cpp
@@ -30,7 +30,7 @@ void QtProjectWizardContentCppStandard::load()
 		std::vector<std::wstring> standards = m_sourceGroupSettings->getAvailableCppStandards();
 		for (size_t i = 0; i < standards.size(); i++)
 		{
-			m_standard->insertItem(i, QString::fromStdWString(standards[i]));
+			m_standard->insertItem(static_cast<int>(i), QString::fromStdWString(standards[i]));
 		}
 
 		m_standard->setCurrentText(QString::fromStdWString(m_sourceGroupSettings->getCppStandard()));
