+++ src/lib_gui/qt/project_wizard/content/paths/QtProjectWizardContentPathsHeaderSearch.cpp
@@ -212,7 +212,7 @@ void QtProjectWizardContentPathsHeaderSearch::validateIncludesButtonClicked()
 						sourceFilePaths,
 						utility::toSet(indexedFilePaths),
 						utility::toSet(headerSearchPaths),
-						log2(sourceFilePaths.size()),
+						static_cast<size_t>(log2(sourceFilePaths.size())),
 						[&](const float progress) {
 							dialogView->showProgressDialog(
 								L"Processing",
@@ -284,7 +284,7 @@ void QtProjectWizardContentPathsHeaderSearch::finishedSelectDetectIncludesRootPa
 						sourceFilePaths,
 						utility::toSet(searchedPaths),
 						utility::toSet(headerSearchPaths),
-						log2(sourceFilePaths.size()),
+						static_cast<size_t>(log2(sourceFilePaths.size())),
 						[&](const float progress) {
 							dialogView->showProgressDialog(
 								L"Processing",
